Adding unit test for detailsByHeight
